UnitCellInspectors work
Created InspectorPanelExtensions, CellActionButtonExtensions, IBuildable.

Made FieldCell and WaterCell implement some interfaces

Refactored a lot of the Inspector functionalities to be handled with OnEnable and OnDisable. Also rewrote inspector toggle functions to pass in the panels I want to toggle so I don't have to rewrite this for future panels.

Also made a bunch of panels as subclasses of InspectorPanel for cleaner SetActive methods.